+++ packages/django-tolap/django_tolap/__init__.py
@@ -8,6 +8,12 @@ from django_tolap.exceptions import TolapDenied, TolapSchemaMismatch, Uninspecta
 from django_tolap.pushdown import EnforcementMode
 from django_tolap.raw import enforce_raw, enforce_sql
 from django_tolap.tool import ToolContext, tolap_context, tolap_tool
+from django_tolap.writes import (
+    enforce_delete,
+    enforce_queryset_delete,
+    enforce_save,
+    enforce_update,
+)
 
 __version__ = version("django-tolap")
 
@@ -19,6 +25,10 @@ __all__ = [
     "Uninspectable",
     "accept_context",
     "enforce",
+    "enforce_delete",
+    "enforce_queryset_delete",
+    "enforce_save",
+    "enforce_update",
     "enforce_raw",
     "enforce_sql",
     "issue_context",